// loot_filter/src/Enums/EndgameItemId.ts
/**
 * Essence item identifiers.
 * Used to create Token of Absolution via Horadric Cube.
 */
export enum EssenceId {
	/** Twisted Essence of Suffering (Andariel/Duriel) */
	TWISTED = 'tes',
	/** Charged Essence of Hatred (Mephisto) */
	CHARGED = 'ceh',
	/** Burning Essence of Terror (Diablo) */
	BURNING = 'bet',
	/** Festering Essence of Destruction (Baal) */
	FESTERING = 'fed',
}

/**
 * Pandemonium Event key identifiers.
 * Used to open portals to Über bosses.
 */
export enum PandemoniumKeyId {
	/** Key of Terror (Countess) */
	TERROR = 'pk1',
	/** Key of Hate (Summoner) */
	HATE = 'pk2',
	/** Key of Destruction (Nihlathak) */
	DESTRUCTION = 'pk3',
}

/**
 * Pandemonium Event organ identifiers.
 * Dropped by mini-übers, used to open Über Tristram.
 */
export enum PandemoniumOrganId {
	/** Diablo's Horn */
	HORN = 'dhn',
	/** Baal's Eye */
	EYE = 'bey',
	/** Mephisto's Brain */
	BRAIN = 'mbr',
}

/**
 * Special endgame item identifiers.
 */
export enum EndgameItemId {
	/** Token of Absolution (respec item) */
	TOKEN = 'toa',
	/** Standard of Heroes (proof of Über Tristram completion) */
	STANDARD = 'std',
}